# Conflicts:
#	src/blitter/32bpp_sse2.hpp
#	src/cheat_gui.cpp
#	src/command_type.h
#	src/company_gui.cpp
#	src/console_cmds.cpp
#	src/core/bitmath_func.hpp
#	src/core/endian_func.hpp
#	src/core/endian_type.hpp
#	src/core/enum_type.hpp
#	src/direction_type.h
#	src/gamelog.cpp
#	src/gfx_type.h
#	src/network/core/tcp.h
#	src/network/core/tcp_game.cpp
#	src/network/core/tcp_game.h
#	src/network/network_client.cpp
#	src/network/network_client.h
#	src/network/network_content_gui.cpp
#	src/network/network_query.h
#	src/network/network_server.h
#	src/newgrf.cpp
#	src/newgrf_commons.h
#	src/newgrf_config.cpp
#	src/newgrf_debug_gui.cpp
#	src/newgrf_roadstop.h
#	src/newgrf_storage.cpp
#	src/openttd.cpp
#	src/saveload/afterload.cpp
#	src/saveload/saveload.cpp
#	src/saveload/saveload.h
#	src/screenshot.cpp
#	src/script/api/script_bridge.hpp
#	src/script/api/script_controller.hpp
#	src/script/api/script_event_types.hpp
#	src/script/api/script_industrytype.hpp
#	src/script/api/script_sign.hpp
#	src/script/api/script_types.hpp
#	src/script/api/script_vehicle.hpp
#	src/settings.cpp
#	src/signal.cpp
#	src/smallmap_gui.cpp
#	src/smallmap_gui.h
#	src/spritecache.h
#	src/strgen/CMakeLists.txt
#	src/table/settings/misc_settings.ini
#	src/timer/timer_game_common.h
#	src/timer/timer_game_tick.h
#	src/toolbar_gui.cpp
#	src/toolbar_gui.h
#	src/vehicle_base.h
#	src/vehicle_gui_base.h
#	src/viewport.cpp
#	src/waypoint_base.h

+++ src/train_cmd.cpp
@@ -5824,7 +5824,7 @@ bool TrainController(Train *v, Vehicle *nomove, bool reverse)
 				Direction chosen_dir = (Direction)b[2];
 
 				/* Call the landscape function and tell it that the vehicle entered the tile */
-				uint32_t r = (v->track & TRACK_BIT_WORMHOLE) ? 0 : VehicleEnterTile(v, gp.new_tile, gp.x, gp.y);
+				uint32_t r = (v->track & TRACK_BIT_WORMHOLE) ? 0 : (uint32_t)VehicleEnterTile(v, gp.new_tile, gp.x, gp.y);
 				if (HasBit(r, VETS_CANNOT_ENTER)) {
 					goto invalid_rail;
 				}
